http => dio
networking => models/user
